requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.1
django-debug-toolbar==4.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
idna==3.7
kombu==5.3.7
oauthlib==3.2.2
pillow==10.3.0
prompt_toolkit==3.0.46
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.0.4
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
